fixed more issues, but still many to go.
