+++ lib/Models/UserDrawing.js
@@ -33,10 +33,11 @@ var Cartesian3 = require('terriajs-cesium/Source/Core/Cartesian3');
  */
 
 /**
+ * For user drawings, which includes lines and/or a polygon
+ *
  * @alias UserDrawing
  * @constructor
  *
- * For user drawings, which includes lines and/or a polygon
  * @param {Object} options Object with the following properties:
  * @param {Terria} options.terria The Terria instance.
  * @param {String} [options.messageHeader="Draw on Map"] Heading for the dialog which pops up when in user drawing mode
